Fixed compilations issues for Clang|GNU compilers. (Support for Linux) Added code coverage for Clang|GNU.

+++ tests/API Tests/MockWebSocket.h
@@ -61,7 +61,7 @@ public:
         bb.append(data.bytes(), data.size());
 
         sentPackets.push_back(std::move(bb));
-        return data.size();
+        return (int)data.size();
     }
 
     bool close() noexcept override
